Move bzlenv to a separate directory
