+++ warn/BUILD.bazel
@@ -9,8 +9,8 @@ go_library(
     importpath = "github.com/bazelbuild/buildtools/warn",
     visibility = ["//visibility:public"],
     deps = [
-        ":bzlenv",
         "//build:go_default_library",
+        "//bzlenv:go_default_library",
         "//edit:go_default_library",
     ],
 )
@@ -26,18 +26,4 @@ go_test(
     deps = [
         "//testutils",
     ],
-)
-
-go_library(
-    name = "bzlenv",
-    srcs = ["bzlenv.go"],
-    importpath = "github.com/bazelbuild/buildtools/warn/bzlenv",
-    deps = ["//build:go_default_library"],
-)
-
-go_test(
-    name = "bzlenv_test",
-    size = "small",
-    srcs = ["bzlenv_test.go"],
-    embed = [":bzlenv"],
 )
